msm: lpm-stats: Print suspend level stats for system level only
Suspend stats should be printed once for the system level. The other
cluster levels do not have a suspend level and suspend stats should not be
printed for these levels. Move the call to print suspend stats to the
overall print stats function to avoid repeated printing of suspend stats
for every cluster level.

Change-Id: I8e2874e479b7a197fc9b45a1b359093e6695a4d4
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>