soc: cpu_pwr_ctl: Fix reg clobbering in 'qcom,cbf-clock-seq' workaround
Commit 9f7f69d2 ("soc: cpu_pwr_ctl: Update 'qcom,cbf-clock-seq'
workaround") introduced the additional steps of saving and restoring
clock mux registers for the local cluster, in addition to the peer
cluster.  Unfortunately, this update failed to take into account that
select_cbf_source() is stateful (has local static variables). As a
result, the second of the two consecutive calls stomped on the saved
state from the first:
	select_cbf_source(local_clksel, false);
	select_cbf_source(peer_clksel, false);

Later, when attempting to restore the muxes to their original position,
the wrong value was restored.  Correct this.

Change-Id: I3aaee71e0fbadbeb878f7e896e300995954c6ab6
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>