msm: clock-local2: Allow checking the clock enable bit during handoff
Currently, during clk handoff, we determine if the clock is ON
in hardware by looking at the status bit. This might not always
work, now with the new votable smmu clock support. The votable
bits could be used to turn on multiple clocks and just looking
at the status bit for the individual clocks might be misleading.
Check the enable bit for the HLOS vote instead for these clocks.

Change-Id: I2ba3f4aed986ef1af17bddb14cac29df74857e7f
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>