soc: qcom: Add cpu_kill functionality for AARCH64 qcom chipsets
Caller CPU v7 architecture has this support of waiting on the callee CPU
before it considers the CPU to be completely asleep. The change has not
been propagated to v8 architecture.

Adding this support to v8 architecture such that we remove the condition
of CPU being woken up even before it is collapsed.

Change-Id: I9c5fb5ca004ecc81c8281705dac616265f9d0721
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>