soc: qcom: mpm: Get irq size for a tree type irq domain
GIC v3 is a tree type irq domain and the revmap_size for the domain is 0
as per the new implementation. This causes NULL pointer issue during mpm
initialization as the driver expects the revmap_size to be always
non-zero.

Assume a max size of 1024 if revmap_size is 0 for any given irq domain.

Change-Id: I22602107bfdbbfbe3432726da230c4566077cd5e
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>