msm: rpm-smd: Fix glink handle race condition
In the current glink framework implementation there is no definite
ordering between when client receieves the CONNECTED notification
and when the glink_open() call returns. In some scenarios this
creates a race condition between the CONNECTED notification and
the glink_open() call. In the current implemenation rpm-smd driver
allows clients to send and receieve messages between APPS and RPM
once it receives a CONNECTED callback. But at this time glink handle
was NULL and it is causing glink_tx() fail.

Avoid race condition by filling the glink handle in CONNECTED
callback.

CRs-fixed: 812580
Change-Id: I103f1d7695804f4aa4d7754989c6684c759ebfbd
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>