regulator: spm-regulator: selectively use SPM API to get regulator voltage
msm_spm_get_vdd() may be used to determine the current regulator
voltage when SPM isn't bypassed. This function call has a lower
latency compared to the SPMI read interface used to find the
regulator voltage at the PMIC. Enable usage of this API when AVS
regulator is in use and fall back to SPMI read if the API is
unavailable.

Change-Id: I4631132af92a5dda10132e05ccb342bbb40faeed
Signed-off-by: Osvaldo Banuelos <osvaldob@codeaurora.org>