Revert "msm: lpm-levels: Initialize last level for cluster at probe"
The original patch was implemented without the PSCI code and worked as
expected on 3.10 kernel. It was added to address an issue where CCI
would enter power collapse when one of the cpus in performance cluster
was online. Uninitialized last_level value of L2 caused CPU that came
online to never unprepare the CCI level (uninitialized value implied
active/default level and implied CCI wasn't doing any low power mode).
Initializing the value to nlevels-1 resolved the issues as this
unprepared CCI when a CPU was brought out of reset.

With PSCI implementation, last_level was initialized to default_level
(active) if all the child cpus within that cluster were online and
nlevel-1 otherwise. This addresses the issue above as well. But
propagating the original change to 3.18, overwrote the initialization
done as part of PSCI. This causes cluster reconfiguration to happen for
M4M the first time even though M4M was never configured for any low
power mode. Revert this change to address this issue.

Change-Id: I34bdc9e618869809f9ad51fed714147ddf142f6b
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>