regulator: rpm-smd-regulator: refactor set_voltage and get_voltage ops
The set_voltage and get_voltage ops for level and corner values are
nearly identical to the ops for plain voltages.

Refactor these ops into a single op for set_voltage and a single op
for get_voltage. Use a new voltage_index struct element to distinguish
which voltage param should be used when these ops are called.

Change-Id: Ibae9164c91fc22dadea1f8d4fb77ec183c3760b9
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>