F2FS: revert flag 0
this was, as it seems to me, carelessly transferred to the comments of f2fs: catch up to 4.4-rc1. The code has been adapted to the old ads in mm.h. Remove the extra flag so as not to cancel comments and not to understand the code that was added upstream to us