Cgroup.h: revert allow_attach
allow_attach was removed in the comment (Revert "cgroup: Add generic c group subsystem permission checks"), but a comment was added (cgroup: Fix issues in allow_attach callback) and the build will not be successful because allow_attach is not declared. Fix it and return the function as it was done in 3.18 magically.